+++ include/boost/beast/websocket/impl/ping.hpp
@@ -12,8 +12,8 @@
 
 #include <boost/beast/core/async_op_base.hpp>
 #include <boost/beast/core/bind_handler.hpp>
+#include <boost/beast/core/stream_traits.hpp>
 #include <boost/beast/core/type_traits.hpp>
-#include <boost/beast/core/detail/get_executor_type.hpp>
 #include <boost/beast/websocket/detail/frame.hpp>
 #include <boost/asio/coroutine.hpp>
 #include <boost/asio/post.hpp>
@@ -33,7 +33,7 @@ template<class NextLayer, bool deflateSupported>
 template<class Handler>
 class stream<NextLayer, deflateSupported>::ping_op
     : public beast::stable_async_op_base<
-        Handler, beast::detail::get_executor_type<stream>>
+        Handler, beast::executor_type<stream>>
     , public net::coroutine
 {
     struct state
@@ -66,7 +66,7 @@ public:
         detail::opcode op,
         ping_data const& payload)
         : stable_async_op_base<
-            Handler, beast::detail::get_executor_type<stream>>(
+            Handler, beast::executor_type<stream>>(
                 std::forward<Handler_>(h), ws.get_executor())
         , d_(beast::allocate_stable<state>(
             *this, ws, op, payload))