+++ include/boost/beast/websocket/impl/read.hpp
@@ -18,11 +18,11 @@
 #include <boost/beast/core/buffers_prefix.hpp>
 #include <boost/beast/core/buffers_suffix.hpp>
 #include <boost/beast/core/flat_static_buffer.hpp>
+#include <boost/beast/core/stream_traits.hpp>
 #include <boost/beast/core/type_traits.hpp>
 #include <boost/beast/core/detail/buffer.hpp>
 #include <boost/beast/core/detail/clamp.hpp>
 #include <boost/beast/core/detail/config.hpp>
-#include <boost/beast/core/detail/get_executor_type.hpp>
 #include <boost/asio/coroutine.hpp>
 #include <boost/asio/post.hpp>
 #include <boost/assert.hpp>
@@ -47,7 +47,7 @@ template<
     class Handler>
 class stream<NextLayer, deflateSupported>::read_some_op
     : public beast::async_op_base<
-        Handler, beast::detail::get_executor_type<stream>>
+        Handler, beast::executor_type<stream>>
     , public net::coroutine
 {
     stream& ws_;
@@ -68,7 +68,7 @@ public:
         stream<NextLayer, deflateSupported>& ws,
         MutableBufferSequence const& bs)
         : async_op_base<
-            Handler, beast::detail::get_executor_type<stream>>(
+            Handler, beast::executor_type<stream>>(
                 std::forward<Handler_>(h), ws.get_executor())
         , ws_(ws)
         , bs_(bs)
@@ -631,7 +631,7 @@ template<
     class Handler>
 class stream<NextLayer, deflateSupported>::read_op
     : public beast::async_op_base<
-        Handler, beast::detail::get_executor_type<stream>>
+        Handler, beast::executor_type<stream>>
     , public net::coroutine
 {
     stream<NextLayer, deflateSupported>& ws_;
@@ -649,7 +649,7 @@ public:
         std::size_t limit,
         bool some)
         : async_op_base<
-            Handler, beast::detail::get_executor_type<stream>>(
+            Handler, beast::executor_type<stream>>(
                 std::forward<Handler_>(h), ws.get_executor())
         , ws_(ws)
         , b_(b)